+++ frontend/src/views/ContentView.svelte
@@ -70,8 +70,8 @@
       "ofrak.core.instruction.Instruction",
       "ofrak.core.data.DataWord",
     ].some((tag) => $selectedResource.has_tag(tag));
-    hasDecompView = ["ofrak.core.decompilation.DecompilationAnalysis"].some(
-      (tag) => $selectedResource.has_tag(tag)
+    hasDecompView = ["ofrak.core.complex_block.ComplexBlock"].some((tag) =>
+      $selectedResource.has_tag(tag)
     );
     if (hasTextView) {
       tabs.push(textTab);